Refactor: change ctrl + c event to be captured only during api call
